--- src/lib/api/apiCollections/index.js
@@ -1,5 +1,0 @@
-export * from "./categories"
-export * from "./location"
-export * from "./orders"
-export * from "./service-request"
-export * from "./services"
