+++ src/lib/api/categories.js
