+++ src/lib/api/index.js
@@ -1,13 +1,17 @@
 import axios from "axios"
-export * from "./apiCollections"
 
+// EXPORT API COLLECTIONS HERE
+export * from "./categories"
+export * from "./location"
+export * from "./orders"
+export * from "./service-request"
+export * from "./services"
+
+// AXIOS INSTANCE EXPORT BY DEFAULT
+// PLEASE DISCUSS BEFORE YOU WANT TO EDIT THIS SCRIPT
 const apiClientRequest = axios.create({
     baseURL: process.env.VUE_APP_BACKEND_API,
-    headers: { "Content-Type": "application/json" },
-    auth: {
-        username: process.env.VUE_APP_USERNAME,
-        password: process.env.VUE_APP_PASSWORD
-    }
+    headers: { "Content-Type": "application/json" }
 })
 
 export default apiClientRequest
