+++ src/store/wallet/index.js
@@ -17,7 +17,11 @@ export default {
       try {
         const debioApiKey = process.env.VUE_APP_DEBIO_API_KEY
         const { data } = await apiClientRequest.post("/substrate/wallet-binding", payload , {
-          headers: { "debio-api-key" : debioApiKey }
+          headers: { "debio-api-key" : debioApiKey },
+          auth: {
+            username: process.env.VUE_APP_USERNAME,
+            password: process.env.VUE_APP_PASSWORD
+          }
         })
 
         commit('SET_RESULT', data)
